pyproject.toml
[project]
name = "code-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "appnope==0.1.4",
    "asttokens==3.0.0",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.1.8",
    "comm==0.2.3",
    "contourpy==1.3.0",
    "cycler==0.12.1",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "exceptiongroup==1.3.0",
    "executing==2.2.1",
    "filelock==3.19.1",
    "fire>=0.7.1",
    "fonttools==4.60.1",
    "fsspec==2025.9.0",
    "hf-xet==1.1.10",
    "huggingface-hub==0.36.0",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "ipykernel==6.31.0",
    "ipython==8.18.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "kiwisolver==1.4.7",
    "markupsafe==3.0.3",
    "matplotlib==3.9.4",
    "matplotlib-inline==0.2.1",
    "mpmath==1.3.0",
    "nest-asyncio==1.6.0",
    "networkx==3.2.1",
    "nltk==3.9.2",
    "numpy==2.0.2",
    "packaging==25.0",
    "pandas==2.3.3",
    "parso==0.8.5",
    "patsy==1.0.2",
    "pexpect==4.9.0",
    "pillow==11.3.0",
    "platformdirs==4.4.0",
    "prompt-toolkit==3.0.52",
    "psutil==7.1.1",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pycparser==2.23",
    "pygments==2.19.2",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "regex==2025.10.23",
    "requests==2.32.5",
    "rpy2==3.6.4",
    "rpy2-rinterface==3.6.3",
    "rpy2-robjects==3.6.3",
    "safetensors==0.6.2",
    "scikit-learn==1.6.1",
    "scipy==1.13.1",
    "seaborn==0.13.2",
    "six==1.17.0",
    "stack-data==0.6.3",
    "statsmodels==0.14.5",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "timm==1.0.20",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers>=4.57.1",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "urllib3==2.5.0",
    "wcwidth==0.2.14",
    "zipp==3.23.0",
]
